+++ config/bundle.js
@@ -1,9 +1,6 @@
-'use strict'
 const fs = require('fs')
-const readFile = fs.readFile
-const writeFile = fs.writeFile
-const relative = require('path').relative
-const gzip = require('zlib').gzip
+const path = require('path')
+const zlib = require('zlib')
 const rollup = require('rollup')
 const uglify = require('uglify-js')
 
@@ -24,49 +21,48 @@ function build (entries) {
 }
 
 function buildEntry (config) {
-  const isProd = /min\.js$/.test(config.dest)
-  return rollup.rollup(config).then(bundle => {
-    const code = bundle.generate(config).code
-    if (isProd) {
-      var minified = (config.banner ? config.banner + '\n' : '') + uglify.minify(code, {
-        fromString: true,
-        output: {
-          screw_ie8: true,
-          ascii_only: true
-        },
-        compress: {
-          pure_funcs: ['makeMap']
-        }
-      }).code
-      return write(config.dest, minified).then(zip(config.dest))
-    } else {
-      return write(config.dest, code)
-    }
-  })
+  const output = config.output
+  const { file, banner } = output
+  const isProd = /min\.js$/.test(file)
+  return rollup.rollup(config)
+    .then(bundle => bundle.generate(output))
+    .then(({ code }) => {
+      if (isProd) {
+        var minified = (banner ? banner + '\n' : '') + uglify.minify(code, {
+          fromString: true,
+          output: {
+            ascii_only: true
+          },
+          compress: {
+            pure_funcs: ['makeMap']
+          }
+        }).code
+        return write(file, minified, true)
+      } else {
+        return write(file, code)
+      }
+    })
 }
 
-function write (dest, code) {
-  return new Promise(function (resolve, reject) {
-    writeFile(dest, code, function (err) {
-      if (err) { return reject(err) }
-      console.log(blue(relative(process.cwd(), dest)) + ' ' + getSize(code))
+function write (dest, code, zip) {
+  return new Promise((resolve, reject) => {
+    function report (extra) {
+      console.log(blue(path.relative(process.cwd(), dest)) + ' ' + getSize(code) + (extra || ''))
       resolve()
+    }
+
+    fs.writeFile(dest, code, err => {
+      if (err) { return reject(err) }
+      if (zip) {
+        zlib.gzip(code, (err, zipped) => {
+          if (err) { return reject(err) }
+          report(' (gzipped: ' + getSize(zipped) + ')')
+        })
+      } else {
+        report()
+      }
     })
   })
-}
-
-function zip (file) {
-  return function () {
-    return new Promise(function (resolve, reject) {
-      readFile(file, function (err, buf) {
-        if (err) { return reject(err) }
-        gzip(buf, function (err, buf) {
-          if (err) { return reject(err) }
-          write(file + '.gz', buf).then(resolve)
-        })
-      })
-    })
-  }
 }
 
 function getSize (code) {