+++ config/webpack.dev.conf.js
@@ -3,6 +3,10 @@ const webpack = require('webpack')
 
 module.exports = {
   entry: 'mocha-loader!./test/unit/index.js',
+  mode: 'development',
+  resolve: {
+    extensions: ['.js']
+  },
   output: {
     path: path.resolve(__dirname, '/test/unit'),
     filename: 'tests.js',
@@ -15,12 +19,5 @@ module.exports = {
       loader: 'babel-loader'
     }]
   },
-  plugins: [
-    new webpack.DefinePlugin({
-      'process.env': {
-        NODE_ENV: '"development"'
-      }
-    })
-  ],
   devtool: '#eval-source-map'
 }